Move PeerWeb rewrite into OmniBox UI
