+++ src/window.js
@@ -13,7 +13,6 @@ import PQueue from 'p-queue'
 import delay from 'delay'
 
 import Config from './config.js'
-import { rewritePeerWebURL } from './url-rewrite.js'
 
 const IS_DEBUG = process.env.NODE_ENV === 'debug'
 
@@ -393,15 +392,6 @@ export class Window extends EventEmitter {
 
     this.window.setBrowserView(this.view)
 
-    this.web.on('will-navigate', (event, url) => {
-      const rewritten = rewritePeerWebURL(url)
-      if (rewritten === url) return
-      event.preventDefault()
-      this.loadURL(rewritten).catch((e) => {
-        console.error('Error rewriting navigation URL', e.stack)
-      })
-    })
-
     this.web.on('did-start-navigation', (event, url, isInPlace, isMainFrame) => {
       this.emitNavigate(url, isMainFrame)
     })
@@ -538,8 +528,7 @@ export class Window extends EventEmitter {
    * @param {string} url
    */
   async loadURL (url) {
-    const rewritten = rewritePeerWebURL(url)
-    return this.web.loadURL(rewritten)
+    return this.web.loadURL(url)
   }
 
   async getURL () {
